+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.2.1-SNAPSHOT
 
 - tailwindcss > /dev/null 2> &1
+- get  admin/new => admin/new
+- post admin/new => admin/create! => admin/upsert!
 
 # 0.2.0 (2025-09-11)